+++ src/Auth/ExternalUser.php
@@ -12,23 +12,56 @@ class ExternalUser extends Authenticatable implements AccessUserInterface
     use HasAccess;
     use Notifiable;
 
-    protected $attributes = [];
+    protected $guarded = [];
+
+    protected array $claims = [];
 
     public function __construct(array $attributes = [])
     {
         parent::__construct([]);
-        $this->attributes = $attributes;
+        $this->fillAttributes($attributes);
+    }
+
+    protected function fillAttributes(array $attributes): void
+    {
+        if (empty($attributes)) {
+            return;
+        }
+
+        $this->claims = $attributes['token_claims'] ?? $attributes;
+
+        if (isset($attributes['token_claims'])) {
+            unset($attributes['token_claims']);
+        }
+
+        $this->forceFill($attributes);
         $this->syncOriginal();
+    }
+
+    public function setClaims(array $claims): void
+    {
+        $this->claims = $claims;
+    }
+
+    public function getClaims(): array
+    {
+        return $this->claims;
     }
 
     public function getAuthIdentifierName()
     {
-        return 'id';
+        return config('microservice.auth.user_identifier_claim', 'id');
     }
 
     public function getAuthIdentifier()
     {
-        return $this->attributes[$this->getAuthIdentifierName()] ?? null;
+        $identifier = parent::getAuthIdentifier();
+
+        if ($identifier === null && $this->getAuthIdentifierName() !== 'sub') {
+            return $this->getAttribute('sub');
+        }
+
+        return $identifier;
     }
 
     public function getAuthPassword()